Add styling to component elements
