+++ src/App.css
@@ -1,0 +1,7 @@
+.main{
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100vh;
+}
