+++ src/components/OrderedList/OrderedList.css
@@ -1,0 +1,61 @@
+.ordered-list-container{
+  height: 400px;
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+  align-items: center;
+}
+
+.modifier-wrapper{
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  align-content: center;
+}
+
+.input-wrapper{
+  padding: 5px;
+}
+
+.input-field{
+  width: 250px;
+  height: 20px;
+  font-size: 16px;
+}
+
+.button-wrapper{
+  height: 20px;
+  padding: 5px;
+}
+
+.button-wrapper:hover{
+  border-bottom: 2px solid black; 
+  cursor: pointer;
+}
+
+.sort-list-button{
+  background-color: #2B78D7;  
+}
+
+
+.clear-list-button{
+  background-color: #F03A18;
+}
+
+.sort-list-button,
+.clear-list-button{
+  color: white;
+  font-size: 16px;
+  font-weight: 700;
+}
+
+
+.list-wrapper{
+  align-self: start;
+}
+
+
+.list-item{
+
+}
