--- src/components/OrderedList1/index.js
@@ -1,1 +1,0 @@
-export { default } from "./OrderedList1"
